requirements.txt
numpy
opencv-python
matplotlib
scikit-image
imageio
torch
torchvision
torchsummary
tensorboardX
pyzmq
pandas
open3d
sklearn